+++ s4e-web/cypress/integration/map/map-layers.spec.ts
@@ -22,9 +22,56 @@ context('Map Layers', () => {
       .selectNthSidebarLayer(0)
       .activeLayersCountShouldBe(0);
   });
+  it('should display url errors', async () => {
+    const label = 'Test';
+    const url = [
+      'https://test-page.pl/?SERVICE=unknown',
+      'REQUEST=unknown',
+      'LAYERS=unkn?own',
+      'STYLES=unkn?own',
+      'FORMAT=image/unknown',
+      'TRANSPARENT=none',
+      'VERSION=x.y.z',
+      'HEIGHT=none',
+      'WIDTH=none',
+      'CRS=unknown',
+      'BBOX=unknown,unknown,unknown,unknown'
+    ].join('&');
+    Layers
+      .openManagementModal()
+      .fillForm(label, url)
+      .errorsCountShouldBe(8);
+  });
+  it('should display information about not recognized params', async () => {
+    const label = 'Test';
+    const url = [
+      'https://test-page.pl/?additional=unknown',
+      'addititonal2=unknown'
+    ].join('&');
+    const initialSize = (await promisify(Layers.pageObject.getSidebarLayers())).length;
+    Layers
+      .openManagementModal()
+      .addNew(label, url)
+      .changeContextTo(ConfirmModal)
+      .contentShouldContain('additional, additional2')
+      .acceptAndChangeContextTo(GeneralModal)
+      .closeAndChangeContext(Layers)
+      .sidebarLayersCountShouldBe(initialSize + 1)
+
+      .openManagementModal()
+      .toggleNthInPanelDisplay(initialSize)
+      .changeContextTo(GeneralModal)
+      .closeAndChangeContext(Layers)
+      .sidebarLayersCountShouldBe(initialSize)
+
+      .openManagementModal()
+      .removeNthWithPermission(0)
+      .changeContextTo(ConfirmModal)
+      .accept();
+  });
   it('should add and remove from panel', async () => {
     const label = 'Test';
-    const url = 'http://test.pl:5000';
+    const url = 'http://localhost:5000/wms';
     const initialSize = (await promisify(Layers.pageObject.getSidebarLayers())).length;
     Layers
       .openManagementModal()
@@ -46,7 +93,7 @@ context('Map Layers', () => {
   });
   it('should add new layer and remove it', async () => {
     const label = 'Test';
-    const url = 'http://test.pl:5000';
+    const url = '/test';
     const initialSize = (await promisify(
       Layers
         .openManagementModal()